+++ labgrid/driver/smallubootdriver.py
@@ -59,7 +59,7 @@ class SmallUBootDriver(UBootDriver):
         # wait until UBoot has reached it's prompt
         self.console.expect(self.prompt)
 
-    def _run(self, cmd: str, *, timeout: int = 30, codec: str = "utf-8", decodeerrors: str = "strict"):  # pylint: disable=unused-argument,line-too-long
+    def _run(self, cmd: str, *, timeout: int = 30):
         """
         If Uboot is in Command-Line mode: Run command cmd and return it's
         output.
@@ -67,7 +67,7 @@ class SmallUBootDriver(UBootDriver):
         Arguments:
         cmd - Command to run
         """
-        # TODO: use codec, decodeerrors
+        # TODO: use self.codec, self.decodeerrors
 
         # Check if Uboot is in command line mode
         if self._status != 1: